requirements.txt
accelerate>=1.1.0,<2.0.0
bitsandbytes>=0.46.0,<0.49.0
datasets>=3.0.0,<4.0.0
huggingface_hub>=0.34.0,<1.0.0
jinja2>=3.1.4,<4.0.0
numpy>=2.1.0,<3.0.0
packaging>=24.1,<26.0
peft>=0.16.0,<0.19.0
pyyaml>=6.0.2,<7.0.0
safetensors>=0.5.0,<1.0.0
sentencepiece>=0.2.0,<1.0.0
transformers>=4.55.0,<4.60.0
trl>=0.20.0,<0.25.0
unsloth>=2026.3.0
vllm>=0.11.0
